Changed toDouble to .0, and added more triangle tests

+++ Assignment1/src/test/kotlin/CircleTest.kt
@@ -2,9 +2,9 @@ import kotlin.test.*
 import kotlin.math.PI
 
 internal class CircleTest {
-    private val testPoint: Point = Point(1.toDouble() ,1.toDouble())
+    private val testPoint: Point = Point(1.0 ,1.0)
 
-    private val testCircle: Circle = Circle(testPoint, 1.toDouble())
+    private val testCircle: Circle = Circle(testPoint, 1.0)
 
     @Test
     fun testArea(){
