+++ Assignment1/src/test/kotlin/EllipseTest.kt
@@ -2,9 +2,9 @@ import kotlin.test.*
 import kotlin.math.PI
 
 internal class EllipseTest {
-    private val testPoint: Point = Point(1.toDouble() ,1.toDouble())
+    private val testPoint: Point = Point(1.0 ,1.0)
 
-    private val testEllipse: Ellipse = Ellipse(testPoint, 1.toDouble(), 1.toDouble())
+    private val testEllipse: Ellipse = Ellipse(testPoint, 1.0, 1.0)
 
     @Test
     fun testArea(){
