+++ Assignment1/src/test/kotlin/PointTest.kt
@@ -1,13 +1,13 @@
 import kotlin.test.*
 
 internal class PointTest {
-    private val testPoint1: Point = Point(0.toDouble(), 0.toDouble())
-    private val testPoint2: Point = Point(1.toDouble(), 1.toDouble())
-    private val testPoint3: Point = Point(1.toDouble(), 4.toDouble())
+    private val testPoint1: Point = Point(0.0, 0.0)
+    private val testPoint2: Point = Point(1.0, 1.0)
+    private val testPoint3: Point = Point(1.0, 4.0)
 
     @Test
     fun testGetPosition(){
-        val expected = Pair(1.toDouble(), 4.toDouble())
+        val expected = Pair(1.0, 4.0)
         assertEquals(expected, testPoint3.getPosition()) 
     }
